requirements.txt
factory-boy==2.11.1
mock==2.0.0
pyhamcrest==1.9.0
djangorestframework
django-cors-headers
django-river @ git+https://github.com/tamhub/django-river@472d6390f322c430b3bf96d7263a174b63939440
django-heroku
psycopg2
gunicorn
faker
djangorestframework-simplejwt